Stop exporting `compose`
We use `compose` for a single HOC test, so these changes make
it a test helper function, that isn't exported / exposed
externally.
