+++ packages/testing/src/index.ts
@@ -10,6 +10,5 @@ export { createClient } from './utils/createClient';
 export { stripSymbols } from './utils/stripSymbols';
 export { wrap } from './utils/wrap';
 export { wait } from './utils/wait';
-export { compose } from './utils/flowRight';
 
 export * from './mocks/types';
